refactor: rename communication component files
