+++ mava/components/jax/communication/gdn_networks.py
